Fix linter warning + run prettier on everything (#3581)
