+++ e2e/.vscode-test.mjs
@@ -2,18 +2,15 @@
  * See: https://code.visualstudio.com/api/working-with-extensions/testing-extension
  */
 
-import { defineConfig } from '@vscode/test-cli';
+import { defineConfig } from "@vscode/test-cli"
 
 export default defineConfig({
-	label: 'integrationTest',
-	files: 'out/suite/**/*.test.js',
-	workspaceFolder: '.',
+	label: "integrationTest",
+	files: "out/suite/**/*.test.js",
+	workspaceFolder: ".",
 	mocha: {
-		ui: 'tdd',
+		ui: "tdd",
 		timeout: 60000,
 	},
-	launchArgs: [
-		'--enable-proposed-api=RooVeterinaryInc.roo-cline',
-		'--disable-extensions'
-	]
-});
+	launchArgs: ["--enable-proposed-api=RooVeterinaryInc.roo-cline", "--disable-extensions"],
+})
