+++ webview-ui/src/components/settings/ModelInfoView.tsx
@@ -73,7 +73,8 @@ export const ModelInfoView = ({
 		),
 		apiProvider === "gemini" && (
 			<span className="italic">
-				{(selectedModelId === "gemini-2.5-pro-preview-03-25" || selectedModelId === "gemini-2.5-pro-preview-05-06")
+				{selectedModelId === "gemini-2.5-pro-preview-03-25" ||
+				selectedModelId === "gemini-2.5-pro-preview-05-06"
 					? t("settings:modelInfo.gemini.billingEstimate")
 					: t("settings:modelInfo.gemini.freeRequests", {
 							count: selectedModelId && selectedModelId.includes("flash") ? 15 : 2,
