+++ webview-ui/src/i18n/locales/nl/history.json
@@ -1,39 +1,39 @@
 {
-  "recentTasks": "Taken",
-  "viewAll": "Alle taken weergeven",
-  "tokens": "Tokens: ↑{{in}} ↓{{out}}",
-  "cache": "Cache: +{{writes}} → {{reads}}",
-  "apiCost": "API-kosten: ${{cost}}",
-  "history": "Geschiedenis",
-  "exitSelectionMode": "Selectiemodus verlaten",
-  "enterSelectionMode": "Selectiemodus starten",
-  "done": "Gereed",
-  "searchPlaceholder": "Geschiedenis doorzoeken...",
-  "newest": "Nieuwste",
-  "oldest": "Oudste",
-  "mostExpensive": "Duurste",
-  "mostTokens": "Meeste tokens",
-  "mostRelevant": "Meest relevant",
-  "deleteTaskTitle": "Taak verwijderen (Shift + Klik om bevestiging over te slaan)",
-  "tokensLabel": "Tokens:",
-  "cacheLabel": "Cache:",
-  "apiCostLabel": "API-kosten:",
-  "copyPrompt": "Prompt kopiëren",
-  "exportTask": "Taak exporteren",
-  "deleteTask": "Taak verwijderen",
-  "deleteTaskMessage": "Weet je zeker dat je deze taak wilt verwijderen? Deze actie kan niet ongedaan worden gemaakt.",
-  "cancel": "Annuleren",
-  "delete": "Verwijderen",
-  "exitSelection": "Selectie verlaten",
-  "selectionMode": "Selectiemodus",
-  "deselectAll": "Alles deselecteren",
-  "selectAll": "Alles selecteren",
-  "selectedItems": "Geselecteerd {{selected}}/{{total}} items",
-  "clearSelection": "Selectie wissen",
-  "deleteSelected": "Geselecteerde verwijderen",
-  "deleteTasks": "Taken verwijderen",
-  "confirmDeleteTasks": "Weet je zeker dat je {{count}} taken wilt verwijderen?",
-  "deleteTasksWarning": "Verwijderde taken kunnen niet worden hersteld. Zorg ervoor dat je wilt doorgaan.",
-  "deleteItems": "Verwijder {{count}} items",
-  "showAllWorkspaces": "Toon taken van alle werkruimtes"
+	"recentTasks": "Taken",
+	"viewAll": "Alle taken weergeven",
+	"tokens": "Tokens: ↑{{in}} ↓{{out}}",
+	"cache": "Cache: +{{writes}} → {{reads}}",
+	"apiCost": "API-kosten: ${{cost}}",
+	"history": "Geschiedenis",
+	"exitSelectionMode": "Selectiemodus verlaten",
+	"enterSelectionMode": "Selectiemodus starten",
+	"done": "Gereed",
+	"searchPlaceholder": "Geschiedenis doorzoeken...",
+	"newest": "Nieuwste",
+	"oldest": "Oudste",
+	"mostExpensive": "Duurste",
+	"mostTokens": "Meeste tokens",
+	"mostRelevant": "Meest relevant",
+	"deleteTaskTitle": "Taak verwijderen (Shift + Klik om bevestiging over te slaan)",
+	"tokensLabel": "Tokens:",
+	"cacheLabel": "Cache:",
+	"apiCostLabel": "API-kosten:",
+	"copyPrompt": "Prompt kopiëren",
+	"exportTask": "Taak exporteren",
+	"deleteTask": "Taak verwijderen",
+	"deleteTaskMessage": "Weet je zeker dat je deze taak wilt verwijderen? Deze actie kan niet ongedaan worden gemaakt.",
+	"cancel": "Annuleren",
+	"delete": "Verwijderen",
+	"exitSelection": "Selectie verlaten",
+	"selectionMode": "Selectiemodus",
+	"deselectAll": "Alles deselecteren",
+	"selectAll": "Alles selecteren",
+	"selectedItems": "Geselecteerd {{selected}}/{{total}} items",
+	"clearSelection": "Selectie wissen",
+	"deleteSelected": "Geselecteerde verwijderen",
+	"deleteTasks": "Taken verwijderen",
+	"confirmDeleteTasks": "Weet je zeker dat je {{count}} taken wilt verwijderen?",
+	"deleteTasksWarning": "Verwijderde taken kunnen niet worden hersteld. Zorg ervoor dat je wilt doorgaan.",
+	"deleteItems": "Verwijder {{count}} items",
+	"showAllWorkspaces": "Toon taken van alle werkruimtes"
 }
