+++ webview-ui/src/stories/Welcome.mdx
@@ -1,4 +1,4 @@
-import { Meta } from "@storybook/blocks";
+import { Meta } from "@storybook/blocks"
 
 <Meta title="Welcome" />
